requirement.txt
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.26.0
opencv-python>=4.6.0
pillow>=9.3.0
numpy>=1.21.0
tqdm>=4.64.0
matplotlib>=3.6.0
tensorboard>=2.10.0
scikit-learn>=1.1.0
pyyaml>=6.0
